requirements.txt
sympy
pillow
pytesseract
numpy
